CONFLICT (content): Merge conflict in openmp/libomptarget/src/private.h
  CONFLICT (content): Merge conflict in openmp/libomptarget/src/omptarget.cpp

  commit 5f1d4d477902a9c058bd5506f17eeab6b7c5b7f5
  Author: Vyacheslav Zakharin <vyacheslav.p.zakharin@intel.com>
  Date:   Tue Jan 26 13:52:53 2021 -0800

      [libomptarget][NFC] Use portable printf format specifiers.

      Differential Revision: https://reviews.llvm.org/D95476

+++ openmp/libomptarget/src/private.h
@@ -136,9 +136,9 @@ static inline void dumpTargetPointerMappings(const ident_t *Loc,
     SourceInfo Info(HostTargetMap.HstPtrName);
 #if INTEL_CUSTOMIZATION
     INFO(OMP_INFOTYPE_ALL, Device.DeviceID,
-         DPxMOD " " DPxMOD " %-8lu %-8ld %s at %s:%d:%d\n",
+         DPxMOD " " DPxMOD " %-8" PRIuPTR " %-8" PRId64 " %s at %s:%d:%d\n",
          DPxPTR(HostTargetMap.HstPtrBegin), DPxPTR(HostTargetMap.TgtPtrBegin),
-         (long unsigned)(HostTargetMap.HstPtrEnd - HostTargetMap.HstPtrBegin),
+         HostTargetMap.HstPtrEnd - HostTargetMap.HstPtrBegin,
          HostTargetMap.getRefCount(), Info.getName(), Info.getFilename(),
          Info.getLine(), Info.getColumn());
 #endif // INTEL_CUSTOMIZATION
@@ -178,7 +178,7 @@ printKernelArguments(const ident_t *Loc, const int64_t DeviceId,
     else
       type = "use_address";
 #if INTEL_CUSTOMIZATION
-    INFO(OMP_INFOTYPE_ALL, DeviceId, "%s(%s)[%ld] %s\n", type,
+    INFO(OMP_INFOTYPE_ALL, DeviceId, "%s(%s)[%" PRId64 "] %s\n", type,
          getNameFromMapping(varName).c_str(), ArgSizes[i], implicit);
 #endif // INTEL_CUSTOMIZATION
